refactor(cli): use debug collection when reporting errors
