+++ packages/cdktf-cli/lib/terraform.ts
@@ -1,4 +1,5 @@
-import { exec } from "../../../lib/util";
+import { Errors } from "./errors";
+import { exec } from "./util";
 
 export const terraformBinaryName =
   process.env.TERRAFORM_BINARY_NAME || "terraform";
@@ -8,4 +9,6 @@ export const terraformVersion = exec(
   {}
 )
   .then((versionString) => JSON.parse(versionString).terraform_version)
-  .catch((err) => `Unknown: Error loading terraform version ${err}`);
+  .catch((err) =>
+    Errors.Usage(`Unknown: Error loading terraform version ${err}`)
+  );
